refactor(tracer): consolidate root span (+10 squashed commits) Squashed commits: [a4f2234] Get the test passing [8276c98] Get allDescendants() working [379b2eb] Big refactor, tests are mostly passing [6d93312] Lerna handles cross-package dependencies [3567e2e] Rough draft of making Span and RootSpan almost identical [eba18f3] Set 'shared' based on the type of span [b56a884] Reformat tests [b1b6dfb] Use parentSpanId if provided [9d4e601] Tests that ensure nested spans are properly translated for Zipkin [585034e] Add tests for nested spans.
