+++ packages/opencensus-core/src/trace/model/no-record/no-record-span.ts
@@ -13,20 +13,71 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+import {noopLogger} from '../../../common/noop-logger';
+import {Logger} from '../../../common/types';
+import {randomSpanId} from '../../../internal/util';
+import * as configTypes from '../../config/types';
 import * as types from '../types';
-import {NoRecordSpanBase} from './no-record-span-base';
 
-/** Implementation for the Span class that does not record trace events. */
-export class NoRecordSpan extends NoRecordSpanBase implements types.Span {
-  /** set isRootSpan = false */
-  readonly isRootSpan = false;
+const STATUS_OK = {
+  code: types.CanonicalCode.OK
+};
 
-  /**
-   * Constructs a new NoRecordSpanImpl instance.
-   */
-  constructor() {
-    super();
+/** Implementation for the SpanBase class that does not record trace events. */
+export class NoRecordSpan implements types.Span {
+  /** Indicates if this span was started */
+  private startedLocal = false;
+  /** Indicates if this span was ended */
+  private endedLocal = false;
+  /** Indicates if this span was forced to end */
+  // @ts-ignore
+  private truncated = false;
+  /** The Span ID of this span */
+  readonly id: string;
+  /** An object to log information to */
+  logger: Logger = noopLogger;
+  /** A set of attributes, each in the format [KEY]:[VALUE] */
+  attributes: types.Attributes = {};
+  /** A text annotation with a set of attributes. */
+  annotations: types.Annotation[] = [];
+  /** An event describing a message sent/received between Spans */
+  messageEvents: types.MessageEvent[] = [];
+  /** Pointers from the current span to another span */
+  links: types.Link[] = [];
+  /** If the parent span is in another process. */
+  remoteParent = false;
+  /** This span's root span.  If it's a root span, it will point to this */
+  root: NoRecordSpan;
+  /** This span's parent. If it's a root span, must be empty */
+  parentSpan?: NoRecordSpan;
+  /** The resource name of the span */
+  name = 'no-record';
+  /** Kind of span. */
+  kind: types.SpanKind = types.SpanKind.UNSPECIFIED;
+  /** A final status for this span */
+  status: types.Status = STATUS_OK;
+  /** Trace Parameters */
+  activeTraceParams: configTypes.TraceParams = {};
+
+  /** The number of dropped attributes. */
+  droppedAttributesCount = 0;
+  /** The number of dropped links. */
+  droppedLinksCount = 0;
+  /** The number of dropped annotations. */
+  droppedAnnotationsCount = 0;
+  /** The number of dropped message events. */
+  droppedMessageEventsCount = 0;
+
+  /** Constructs a new SpanBaseModel instance. */
+  constructor(parent?: NoRecordSpan) {
+    this.id = randomSpanId();
+    if (parent) {
+      this.root = parent.root;
+      this.parentSpan = parent;
+    } else {
+      this.root = this;
+    }
+    this.logger = (this.root && this.root.logger) || this.logger;
   }
 
   /** Gets trace id of no-record span. */
@@ -34,13 +85,125 @@ export class NoRecordSpan extends NoRecordSpanBase implements types.Span {
     return '';
   }
 
-  get traceState(): types.TraceState {
-    return '';
+  /** Gets the trace state */
+  get traceState(): types.TraceState|undefined {
+    return this.root.traceState;
+  }
+
+  /** Gets the ID of the parent span. */
+  get parentSpanId(): string {
+    if (!this.parentSpan) {
+      return 'no-parent';
+    }
+    return this.parentSpan.id;
+  }
+
+  /** Indicates if span was started. */
+  get started(): boolean {
+    return this.startedLocal;
+  }
+
+  /** Indicates if span was ended. */
+  get ended(): boolean {
+    return this.endedLocal;
   }
 
   /** No-op implementation of this method. */
-  start() {}
+  get startTime(): Date {
+    return new Date();
+  }
 
   /** No-op implementation of this method. */
-  end() {}
+  allDescendants(): types.Span[] {
+    return [];
+  }
+
+  /** No-op implementation of this method. */
+  get spans(): types.Span[] {
+    return [];
+  }
+
+  /** No-op implementation of this method. */
+  get numberOfChildren(): number {
+    return 0;
+  }
+
+  /** No-op implementation of this method. */
+  get endTime(): Date {
+    return new Date();
+  }
+
+  /** Gives the TraceContext of the span. */
+  get spanContext(): types.SpanContext {
+    return {
+      traceId: this.traceId,
+      spanId: this.id,
+      options: 0,
+      traceState: this.traceState
+    };
+  }
+
+  /** No-op implementation of this method. */
+  get duration(): number {
+    return 0;
+  }
+
+  /** No-op implementation of this method. */
+  addAttribute(key: string, value: string|number|boolean) {}
+
+  /** No-op implementation of this method. */
+  addAnnotation(
+      description: string, attributes?: types.Attributes, timestamp = 0) {}
+
+  /** No-op implementation of this method. */
+  addLink(
+      traceId: string, spanId: string, type: types.LinkType,
+      attributes?: types.Attributes) {}
+
+  /** No-op implementation of this method. */
+  addMessageEvent(
+      type: types.MessageEventType, id: number, timestamp = 0,
+      uncompressedSize?: number, compressedSize?: number) {}
+
+  /** No-op implementation of this method. */
+  setStatus(code: types.CanonicalCode, message?: string) {}
+
+  /** No-op implementation of this method. */
+  start() {
+    this.startedLocal = true;
+  }
+
+  /** No-op implementation of this method. */
+  end(): void {
+    this.startedLocal = false;
+    this.endedLocal = true;
+  }
+
+  /** No-op implementation of this method. */
+  truncate() {}
+
+  /**
+   * Starts a new no record child span in the no record root span.
+   * @param nameOrOptions Span name string or SpanOptions object.
+   * @param kind Span kind if not using SpanOptions object.
+   */
+  startChildSpan(
+      nameOrOptions?: string|types.SpanOptions,
+      kind?: types.SpanKind): types.Span {
+    const noRecordChild = new NoRecordSpan(this);
+
+    const spanName =
+        typeof nameOrOptions === 'object' ? nameOrOptions.name : nameOrOptions;
+    const spanKind =
+        typeof nameOrOptions === 'object' ? nameOrOptions.kind : kind;
+    if (spanName) {
+      noRecordChild.name = spanName;
+    }
+    if (spanKind) {
+      noRecordChild.kind = spanKind;
+    }
+
+    noRecordChild.start();
+    return noRecordChild;
+  }
 }
