+++ packages/opencensus-exporter-zipkin/test/test-zipkin.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CanonicalCode, CoreTracer, MessageEventType, RootSpan, SpanKind, TracerConfig} from '@opencensus/core';
+import {CanonicalCode, CoreTracer, MessageEventType, Span, SpanKind, TracerConfig} from '@opencensus/core';
 import * as assert from 'assert';
 import * as nock from 'nock';
 
@@ -72,7 +72,7 @@ describe('Zipkin Exporter', function() {
       tracer.registerSpanEventListener(exporter);
       tracer.start(defaultConfig);
 
-      tracer.startRootSpan({name: 'root-test'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'root-test'}, (rootSpan: Span) => {
         const span =
             rootSpan.startChildSpan({name: 'spanTest', kind: SpanKind.CLIENT});
         span.end();
@@ -90,7 +90,7 @@ describe('Zipkin Exporter', function() {
       tracer.start(defaultConfig);
 
       return tracer.startRootSpan(
-          {name: 'root-test'}, async (rootSpan: RootSpan) => {
+          {name: 'root-test'}, async (rootSpan: Span) => {
             const span = rootSpan.startChildSpan(
                 {name: 'spanTest', kind: SpanKind.CLIENT});
             span.end();
@@ -108,7 +108,7 @@ describe('Zipkin Exporter', function() {
       const tracer = new CoreTracer();
       tracer.start(defaultConfig);
 
-      return tracer.startRootSpan({name: 'root-test'}, (rootSpan: RootSpan) => {
+      return tracer.startRootSpan({name: 'root-test'}, (rootSpan: Span) => {
         const span =
             rootSpan.startChildSpan({name: 'spanTest', kind: SpanKind.CLIENT});
         span.addAttribute('my-int-attribute', 100);
@@ -184,7 +184,7 @@ describe('Zipkin Exporter', function() {
          tracer.start(defaultConfig);
 
          return tracer.startRootSpan(
-             {name: 'root-test'}, async (rootSpan: RootSpan) => {
+             {name: 'root-test'}, async (rootSpan: Span) => {
                const span = rootSpan.startChildSpan(
                    {name: 'spanTest', kind: SpanKind.CLIENT});
                span.end();
