+++ packages/opencensus-instrumentation-grpc/test/test-grpc.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CoreTracer, globalStats, Measurement, RootSpan, Span, SpanEventListener, SpanKind, StatsEventListener, TagKey, TagMap, TagValue, View} from '@opencensus/core';
+import {CoreTracer, globalStats, Measurement, Span, SpanEventListener, SpanKind, StatsEventListener, TagKey, TagMap, TagValue, View} from '@opencensus/core';
 import {logger} from '@opencensus/core';
 import * as assert from 'assert';
 import * as grpcModule from 'grpc';
@@ -251,10 +251,10 @@ const grpcClient = {
 };
 
 class RootSpanVerifier implements SpanEventListener {
-  endedRootSpans: RootSpan[] = [];
+  endedRootSpans: Span[] = [];
 
-  onStartSpan(span: RootSpan): void {}
-  onEndSpan(root: RootSpan) {
+  onStartSpan(span: Span): void {}
+  onEndSpan(root: Span) {
     this.endedRootSpans.push(root);
   }
 }
@@ -496,8 +496,8 @@ describe('GrpcPlugin() ', function() {
                assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
                const options = {name: 'TestRootSpan'};
                const spanName = `grpc.pkg_test.GrpcTester/${method.methodName}`;
-               let serverRoot: RootSpan;
-               return tracer.startRootSpan(options, async (root: RootSpan) => {
+               let serverRoot: Span;
+               return tracer.startRootSpan(options, async (root: Span) => {
                  assert.strictEqual(root.name, options.name);
                  const args = [client, method.request];
                  await method.method.apply(this, args)
@@ -531,42 +531,39 @@ describe('GrpcPlugin() ', function() {
                assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
                const options = {name: 'TestRootSpan'};
                const spanName = `grpc.pkg_test.GrpcTester/${method.methodName}`;
-               let serverRoot: RootSpan;
+               let serverRoot: Span;
                const tags = new TagMap();
                tags.set({name: 'testKey1'}, {value: 'value1'});
                tags.set({name: 'testKey2'}, {value: 'value2'});
                return globalStats.withTagContext(tags, async () => {
-                 return tracer.startRootSpan(
-                     options, async (root: RootSpan) => {
-                       assert.strictEqual(root.name, options.name);
-                       const args = [client, method.request];
-                       await method.method.apply(this, args)
-                           .then(
-                               (result: TestRequestResponse|
-                                TestRequestResponse[]) => {
-                                 assert.ok(checkEqual(result)(method.result));
-                                 assert.strictEqual(
-                                     rootSpanVerifier.endedRootSpans.length, 1);
+                 return tracer.startRootSpan(options, async (root: Span) => {
+                   assert.strictEqual(root.name, options.name);
+                   const args = [client, method.request];
+                   await method.method.apply(this, args)
+                       .then(
+                           (result: TestRequestResponse|
+                            TestRequestResponse[]) => {
+                             assert.ok(checkEqual(result)(method.result));
+                             assert.strictEqual(
+                                 rootSpanVerifier.endedRootSpans.length, 1);
 
-                                 serverRoot =
-                                     rootSpanVerifier.endedRootSpans[0];
-                                 assertSpan(
-                                     serverRoot, spanName, SpanKind.SERVER,
-                                     grpcModule.status.OK);
-                               });
-                       root.end();
-                       assert.strictEqual(
-                           rootSpanVerifier.endedRootSpans.length, 2);
-                       const clientChild =
-                           rootSpanVerifier.endedRootSpans[1].spans[0];
-                       assertSpan(
-                           clientChild, spanName, SpanKind.CLIENT,
-                           grpcModule.status.OK);
-                       // propagation
-                       assertPropagation(clientChild, serverRoot);
-                       assert.deepEqual(
-                           globalStats.getCurrentTagContext(), tags);
-                     });
+                             serverRoot = rootSpanVerifier.endedRootSpans[0];
+                             assertSpan(
+                                 serverRoot, spanName, SpanKind.SERVER,
+                                 grpcModule.status.OK);
+                           });
+                   root.end();
+                   assert.strictEqual(
+                       rootSpanVerifier.endedRootSpans.length, 2);
+                   const clientChild =
+                       rootSpanVerifier.endedRootSpans[1].spans[0];
+                   assertSpan(
+                       clientChild, spanName, SpanKind.CLIENT,
+                       grpcModule.status.OK);
+                   // propagation
+                   assertPropagation(clientChild, serverRoot);
+                   assert.deepEqual(globalStats.getCurrentTagContext(), tags);
+                 });
                });
              });
         });
@@ -618,8 +615,8 @@ describe('GrpcPlugin() ', function() {
                assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
                const options = {name: 'TestRootSpan'};
                const spanName = `grpc.pkg_test.GrpcTester/${method.methodName}`;
-               let serverRoot: RootSpan;
-               return tracer.startRootSpan(options, async (root: RootSpan) => {
+               let serverRoot: Span;
+               return tracer.startRootSpan(options, async (root: Span) => {
                  assert.strictEqual(root.name, options.name);
                  const errRequest = (method.request instanceof Array) ?
                      method.request.slice(0, method.request.length) :
