+++ packages/opencensus-instrumentation-https/test/test-https.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CoreTracer, logger, RootSpan, Span, SpanEventListener} from '@opencensus/core';
+import {CoreTracer, logger, Span, SpanEventListener} from '@opencensus/core';
 import * as assert from 'assert';
 import * as fs from 'fs';
 import * as https from 'https';
@@ -64,10 +64,10 @@ const httpRequest = {
 const VERSION = process.versions.node;
 
 class RootSpanVerifier implements SpanEventListener {
-  endedRootSpans: RootSpan[] = [];
+  endedRootSpans: Span[] = [];
 
-  onStartSpan(span: RootSpan): void {}
-  onEndSpan(root: RootSpan) {
+  onStartSpan(span: Span): void {}
+  onEndSpan(root: Span) {
     this.endedRootSpans.push(root);
   }
 }
@@ -201,7 +201,7 @@ describe('HttpsPlugin', () => {
           const testPath = '/outgoing/rootSpan/childs/1';
           doNock(urlHost, testPath, 200, 'Ok');
           const options = {name: 'TestRootSpan'};
-          return tracer.startRootSpan(options, async (root: RootSpan) => {
+          return tracer.startRootSpan(options, async (root: Span) => {
             await requestMethod(`${urlHost}${testPath}`).then((result) => {
               assert.ok(root.name.indexOf('TestRootSpan') >= 0);
               assert.strictEqual(root.spans.length, 1);
@@ -222,7 +222,7 @@ describe('HttpsPlugin', () => {
                    urlHost, testPath, httpErrorCodes[i],
                    httpErrorCodes[i].toString());
                const options = {name: 'TestRootSpan'};
-               return tracer.startRootSpan(options, async (root: RootSpan) => {
+               return tracer.startRootSpan(options, async (root: Span) => {
                  await requestMethod(`${urlHost}${testPath}`).then((result) => {
                    assert.ok(root.name.indexOf('TestRootSpan') >= 0);
                    assert.strictEqual(root.spans.length, 1);
@@ -242,7 +242,7 @@ describe('HttpsPlugin', () => {
           const num = 5;
           doNock(urlHost, testPath, 200, 'Ok', num);
           const options = {name: 'TestRootSpan'};
-          return tracer.startRootSpan(options, async (root: RootSpan) => {
+          return tracer.startRootSpan(options, async (root: Span) => {
             assert.ok(root.name.indexOf('TestRootSpan') >= 0);
             for (let i = 0; i < num; i++) {
               await requestMethod(`${urlHost}${testPath}`).then((result) => {
