+++ packages/opencensus-instrumentation-ioredis/test/test-ioredis.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CoreTracer, logger, RootSpan, Span, SpanEventListener, SpanKind} from '@opencensus/core';
+import {CoreTracer, logger, Span, SpanEventListener, SpanKind} from '@opencensus/core';
 import * as assert from 'assert';
 import * as ioredis from 'ioredis';
 
@@ -22,12 +22,12 @@ import {plugin} from '../src';
 
 /** Collects ended root spans to allow for later analysis. */
 class RootSpanVerifier implements SpanEventListener {
-  endedRootSpans: RootSpan[] = [];
+  endedRootSpans: Span[] = [];
 
-  onStartSpan(span: RootSpan): void {
+  onStartSpan(span: Span): void {
     return;
   }
-  onEndSpan(root: RootSpan) {
+  onEndSpan(root: Span) {
     this.endedRootSpans.push(root);
   }
 }
@@ -107,7 +107,7 @@ describe('IORedisPlugin', () => {
 
   describe('Instrumenting query operations', () => {
     it('should create a child span for hset', (done) => {
-      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: Span) => {
         client.hset('hash', 'random', 'random', (err, result) => {
           assert.ifError(err);
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
@@ -122,7 +122,7 @@ describe('IORedisPlugin', () => {
     });
 
     it('should create a child span for get', (done) => {
-      tracer.startRootSpan({name: 'getRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'getRootSpan'}, (rootSpan: Span) => {
         client.get('test', (err, result) => {
           assert.ifError(err);
           assert.strictEqual(result, 'data');
@@ -137,17 +137,15 @@ describe('IORedisPlugin', () => {
     });
 
     it('should create a child span for del', (done) => {
-      tracer.startRootSpan(
-          {name: 'removeRootSpan'}, async (rootSpan: RootSpan) => {
-            await client.del('test');
-            assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
-            rootSpan.end();
-            assertSpan(
-                rootSpanVerifier, `redis-del`, SpanKind.CLIENT, (span) => {
-                  return span.attributes.arguments === undefined;
-                });
-            done();
-          });
+      tracer.startRootSpan({name: 'removeRootSpan'}, async (rootSpan: Span) => {
+        await client.del('test');
+        assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
+        rootSpan.end();
+        assertSpan(rootSpanVerifier, `redis-del`, SpanKind.CLIENT, (span) => {
+          return span.attributes.arguments === undefined;
+        });
+        done();
+      });
     });
   });
 
@@ -157,7 +155,7 @@ describe('IORedisPlugin', () => {
     });
 
     it('should not create a child span for insert', (done) => {
-      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: Span) => {
         client.hset('hash', 'random', 'random', (err, result) => {
           assert.ifError(err);
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
@@ -171,7 +169,7 @@ describe('IORedisPlugin', () => {
     });
 
     it('should not create a child span for get', (done) => {
-      tracer.startRootSpan({name: 'getRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'getRootSpan'}, (rootSpan: Span) => {
         client.get('test', (err, result) => {
           assert.ifError(err);
           assert.strictEqual(result, 'data');
@@ -185,16 +183,14 @@ describe('IORedisPlugin', () => {
     });
 
     it('should not create a child span for del', (done) => {
-      tracer.startRootSpan(
-          {name: 'removeRootSpan'}, async (rootSpan: RootSpan) => {
-            await client.del('test');
-            assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
-            rootSpan.end();
-            assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 1);
-            assert.strictEqual(
-                rootSpanVerifier.endedRootSpans[0].spans.length, 0);
-            done();
-          });
+      tracer.startRootSpan({name: 'removeRootSpan'}, async (rootSpan: Span) => {
+        await client.del('test');
+        assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
+        rootSpan.end();
+        assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 1);
+        assert.strictEqual(rootSpanVerifier.endedRootSpans[0].spans.length, 0);
+        done();
+      });
     });
   });
 });
