+++ packages/opencensus-instrumentation-mongodb/test/test-mongodb.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CoreTracer, RootSpan, SpanEventListener, SpanKind} from '@opencensus/core';
+import {CoreTracer, Span, SpanEventListener, SpanKind} from '@opencensus/core';
 import * as assert from 'assert';
 import * as mongodb from 'mongodb';
 import {plugin} from '../src/';
@@ -26,10 +26,10 @@ export type MongoDBAccess = {
 
 /** Collects ended root spans to allow for later analysis. */
 class RootSpanVerifier implements SpanEventListener {
-  endedRootSpans: RootSpan[] = [];
+  endedRootSpans: Span[] = [];
 
-  onStartSpan(span: RootSpan): void {}
-  onEndSpan(root: RootSpan) {
+  onStartSpan(span: Span): void {}
+  onEndSpan(root: Span) {
     this.endedRootSpans.push(root);
   }
 }
@@ -148,7 +148,7 @@ describe('MongoDBPlugin', () => {
     it('should create a child span for insert', (done) => {
       const insertData = [{a: 1}, {a: 2}, {a: 3}];
 
-      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: Span) => {
         collection.insertMany(insertData, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -162,7 +162,7 @@ describe('MongoDBPlugin', () => {
     });
 
     it('should create a child span for update', (done) => {
-      tracer.startRootSpan({name: 'updateRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'updateRootSpan'}, (rootSpan: Span) => {
         collection.updateOne({a: 2}, {$set: {b: 1}}, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -176,7 +176,7 @@ describe('MongoDBPlugin', () => {
     });
 
     it('should create a child span for remove', (done) => {
-      tracer.startRootSpan({name: 'removeRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'removeRootSpan'}, (rootSpan: Span) => {
         collection.deleteOne({a: 3}, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -193,7 +193,7 @@ describe('MongoDBPlugin', () => {
   /** Should intercept cursor */
   describe('Instrumenting cursor operations', () => {
     it('should create a child span for find', (done) => {
-      tracer.startRootSpan({name: 'findRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'findRootSpan'}, (rootSpan: Span) => {
         collection.find({}).toArray((err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -210,7 +210,7 @@ describe('MongoDBPlugin', () => {
   /** Should intercept command */
   describe('Instrumenting command operations', () => {
     it('should create a child span for create index', (done) => {
-      tracer.startRootSpan({name: 'indexRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'indexRootSpan'}, (rootSpan: Span) => {
         collection.createIndex({a: 1}, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -224,7 +224,7 @@ describe('MongoDBPlugin', () => {
     });
 
     it('should create a child span for count', (done) => {
-      tracer.startRootSpan({name: 'countRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'countRootSpan'}, (rootSpan: Span) => {
         collection.count({a: 1}, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -246,7 +246,7 @@ describe('MongoDBPlugin', () => {
     it('should not create a child span for query', (done) => {
       const insertData = [{a: 1}, {a: 2}, {a: 3}];
 
-      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'insertRootSpan'}, (rootSpan: Span) => {
         collection.insertMany(insertData, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -260,7 +260,7 @@ describe('MongoDBPlugin', () => {
     });
 
     it('should not create a child span for cursor', (done) => {
-      tracer.startRootSpan({name: 'findRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'findRootSpan'}, (rootSpan: Span) => {
         collection.find({}).toArray((err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
@@ -274,7 +274,7 @@ describe('MongoDBPlugin', () => {
     });
 
     it('should not create a child span for command', (done) => {
-      tracer.startRootSpan({name: 'indexRootSpan'}, (rootSpan: RootSpan) => {
+      tracer.startRootSpan({name: 'indexRootSpan'}, (rootSpan: Span) => {
         collection.createIndex({a: 1}, (err, result) => {
           assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);
           rootSpan.end();
