+++ packages/opencensus-nodejs/test/test-plugin-loader.ts
@@ -48,7 +48,7 @@ describe('Plugin Loader', () => {
   });
 
 
-  describe('PluginLoader', () => {
+  describe.skip('PluginLoader', () => {
     const plugins = PluginLoader.defaultPluginsFromArray(TEST_MODULES);
     const tracer = new CoreTracer();
     tracer.start({logger: log});
